Fix leaking C++->Python cast of TransformStamped.header.stamp
The manually call to our custom type_caster is missing an ownership
transfer, so the refcount is wrong. Stealing would be correct in this case:

```c++
tf.attr("header").attr("stamp") =
    ::pybind11::reinterpret_steal<object>(
        type_caster<ros::Time>().cast(src.header.stamp, policy, parent));
```

Overall, this is not necessary because `pybind11::cast` automatically
takes care of this and we can just use that instead.

Closes: #16