Shield requests with a mutex
This is part of the C++ client but wasn't ported to the pure Python
client when I removed the bindings (I thought it's not needed).

Turns out that this could lead to regression when users do "special" stuff
like putting the client into a singleton that gets accessed by multiple
threads.